bonus

+++ philo_bonus/srcs_process_semaphore/processes.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   processes.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: node <node@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: ysonmez <ysonmez@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/02 12:58:19 by ysonmez           #+#    #+#             */
-/*   Updated: 2021/12/18 17:09:20 by node             ###   ########.fr       */
+/*   Updated: 2021/12/20 16:56:28 by ysonmez          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,28 +18,41 @@
 *	Wait for child process
 */
 
-static int	fill_philo(t_ph *ph, t_param *param)
+static int	fill_philo(t_param *param)
 {
-	int	i;
+	t_ph	ph;
+	int		i;
+	int		status;
+	pid_t	pid;
 
 	i = 0;
 	while (i < param->nb_philo)
 	{
-		ph[i].param = param;
-		ph[i].i = i + 1;
-		ph[i].last_meal = 0;
-		ph[i].philo = fork();
-		if (ph[i].philo == -1)
+		ph.param = param;
+		ph.meal = 0;
+		ph.i = i + 1;
+		ph.last_meal = 0;
+		ph.philo = fork();
+		if (ph.philo == -1)
 			exit(EXIT_FAILURE);
-		if (ph[i].philo == 0)
-		{
-			schedule(&ph[i]);
-			exit(EXIT_SUCCESS);
-		}
+		if (ph.philo == 0)
+			schedule(&ph);
 		i++;
 	}
-	waitpid(-1, NULL, 0);
-	kill(0, SIGINT);
+	waitpid(-1, &status, WEXITSTATUS(status));
+	{
+		if (status == 256)
+			kill(0, SIGINT);
+		while ((pid = wait(&status)) > 0);
+		kill(0, SIGINT);
+	}
+	//if (status == 256)
+	//	kill(0, SIGINT);
+	//else if (status == 0)
+	//{
+	//	waitpid(-1, NULL, 0);
+	//	kill(0, SIGINT);
+	//}
 	return (0);
 }
 
@@ -47,30 +60,26 @@ static int	fill_philo(t_ph *ph, t_param *param)
 *	and semaphores
 */
 
-t_ph	*create_philo(t_param *param)
+void	create_philo(t_param *param)
 {
-	t_ph	*ph;
-
-	ph = (t_ph *)malloc(sizeof(t_ph) * (param->nb_philo));
-	if (ph == NULL)
-		return (NULL);
-	if (sem_unlink("/forks"))
-		return (NULL);
-	if (sem_unlink("/print"))
-		return (NULL);
-	if (sem_unlink("/end"))
-		return (NULL);
+	sem_unlink("/forks");
+	sem_unlink("/end");
+	sem_unlink("/print");
 	param->forks = sem_open("/forks", O_CREAT, 0660, param->nb_philo);
 	if (param->forks == NULL)
-		return (NULL);
+		exit(EXIT_FAILURE);
 	param->print = sem_open("/print", O_CREAT, 0660, 1);
 	if (param->print == NULL)
-		return (NULL);
+		exit(EXIT_FAILURE);
 	param->end = sem_open("/end", O_CREAT, 0660, 1);
 	if (param->end == NULL)
-		return (NULL);
-	param->alive = true;
-	if (fill_philo(ph, param))
-		return (NULL);
-	return (ph);
+		exit(EXIT_FAILURE);
+	if (fill_philo(param))
+		exit(EXIT_FAILURE);
+	if (sem_unlink("/forks"))
+		exit(EXIT_FAILURE);
+	if (sem_unlink("/print"))
+		exit(EXIT_FAILURE);
+	if (sem_unlink("/end"))
+		exit(EXIT_FAILURE);
 }
